requirements.txt
torch
torchvision
clip
opencv-python
scikit-learn
Pillow
tqdm
matplotlib
